+++ src/components/DataDictionary.tsx
@@ -24,7 +24,7 @@
 
  */
 
-import React, { PureComponent, useState, useEffect, useContext } from "react";
+import React from "react";
 import {
   Chip,
   Flex,
@@ -32,29 +32,22 @@ import {
   Heading,
   Spinner,
   theme,
-  Icon,
 } from "@looker/components";
 import humanize from 'humanize-string'
 import styled, { ThemeProvider } from "styled-components";
-import { useAllModels } from "../utils/fetchers";
+import { useAllModels, getComments } from "../utils/fetchers";
 import "./styles.css";
 import { PanelFields } from "./PanelFields";
 import SidebarToggle from "./SidebarToggle";
 import { useCurrentModel, useCurrentExplore } from "../utils/routes"
-import { ColumnDescriptor, FieldComments } from "./interfaces";
+import { ColumnDescriptor } from "./interfaces";
 import { SQLSnippet } from "./SQLSnippet";
 import { Sidebar } from './Sidebar'
 import { CommentIcon } from './CommentIcon'
 import { SidebarStyleProps } from "./interfaces";
 import { NoModelsAvailable } from "./NoModelsAvailable";
-import { ILookmlModelExploreField, IUser } from "@looker/sdk";
+import { ILookmlModelExploreField } from "@looker/sdk";
 import { CategorizedLabel } from './CategorizedLabel'
-import { getAuthorIds, getMyUser, loadUsers } from "../utils/fetchers";
-import {
-  ExtensionContext,
-  ExtensionContextData,
-} from '@looker/extension-sdk-react'
-import * as semver from 'semver'
 
 export const columns: ColumnDescriptor[] = [
   {
@@ -155,116 +148,7 @@ export const DataDictionary: React.FC<{}> = () => {
   const [sidebarOpen, setSidebarOpen] = React.useState(true)
   const [search, setSearch] = React.useState('')
   const { currentExplore, loadingExplore } = useCurrentExplore()
-  const extensionContext = useContext<ExtensionContextData>(ExtensionContext)
-  const { extensionSDK, initializeError } = extensionContext
-  const [canPersistContextData, setCanPersistContextData] = useState<boolean>(false)
-  const { coreSDK } = useContext(ExtensionContext)
-  const [contextCopy, setContextCopy] = React.useState("{}")
-  const [authors, setAuthors] = React.useState<IUser[]>([])
-  const [comments, setComments] = React.useState("{}")
-  const [me, setMe] = React.useState<IUser>()
-
-
-  const addComment = async (newCommentStr: string, field: string): Promise<boolean> => {
-    let revivedComments = JSON.parse(contextCopy)
-    let newComment = JSON.parse(newCommentStr)
-    
-    if (revivedComments[currentExplore.name][field]) {
-      let revivedFields = revivedComments[currentExplore.name][field]
-      revivedFields.push(newComment)
-      // revivedComments[currentExplore.name][field] = revivedFields
-    } else {
-      revivedComments[currentExplore.name][field] = [newComment]
-    }
-
-    if (canPersistContextData) {
-      try {
-        setComments(JSON.stringify(revivedComments))
-        setContextCopy(JSON.stringify(revivedComments))
-        await extensionSDK.saveContextData(JSON.stringify(revivedComments))
-        return true
-      } catch (error) {
-        console.log(error)
-      }
-    }
-    return false
-  }
-
-  const editComment = async (modCommentStr: string, field: string): Promise<boolean> => {
-    let revivedComments = JSON.parse(contextCopy)
-    let modComment = JSON.parse(modCommentStr)
-
-    let newFieldComments = revivedComments[currentExplore.name][field].filter((d: FieldComments) => {
-      return d.pk !== modComment.pk
-    })
-    newFieldComments.push(modComment)
-    revivedComments[currentExplore.name][field] = newFieldComments
-
-    if (canPersistContextData) {
-      try {
-        setComments(JSON.stringify(revivedComments))
-        setContextCopy(JSON.stringify(revivedComments))
-        await extensionSDK.saveContextData(JSON.stringify(revivedComments))
-        return true
-      } catch (error) {
-        console.log(error)
-      }
-    }
-    return false
-  }
-
-  const deleteComment = async (delCommentStr: string, field: string): Promise<boolean> => {
-    let revivedComments = JSON.parse(contextCopy)
-    let delComment = JSON.parse(delCommentStr)
-
-    let newFieldComments = revivedComments[currentExplore.name][field].filter((d: any) => {
-      return d.pk !== delComment.pk
-    })
-    revivedComments[currentExplore.name][field] = newFieldComments
-
-    if (canPersistContextData) {
-      try {
-        setComments(JSON.stringify(revivedComments))
-        setContextCopy(JSON.stringify(revivedComments))
-        await extensionSDK.saveContextData(JSON.stringify(revivedComments))
-        return true
-      } catch (error) {
-        console.log(error)
-      }
-    }
-    return false
-  }
-  
-  useEffect(() => {
-    const initialize = async () => {
-      // Context requires Looker version 7.14.0. If not supported provide
-      // default configuration object and disable saving of context data.
-      let context
-      if (
-        semver.intersects(
-          '>=7.14.0',
-          extensionSDK.lookerHostData?.lookerVersion || '7.0.0',
-          true
-        )
-      ) {
-        try {
-          context = await extensionSDK.getContextData()
-          setCanPersistContextData(true)
-          setContextCopy(context)
-          let authorIds = getAuthorIds(context)
-          let contextObj = JSON.parse(context)
-          currentExplore && contextObj[currentExplore.name] ? null : contextObj[currentExplore.name] = {}
-          setContextCopy(JSON.stringify(contextObj))
-          setComments(JSON.stringify(contextObj))
-          setAuthors(await loadUsers(coreSDK, authorIds))
-          setMe(await getMyUser(coreSDK))
-        } catch (error) {
-          console.error(error)
-        }
-      }
-    }
-    initialize()
-  }, [currentExplore])
+  const { comments, authors, me, addComment, editComment, deleteComment } = getComments(currentExplore)
 
   let models
